Feature: spam detection on posts and comments
Together with a small service, every new post or comment with a language tag of 'en' will be tested if it is spam or not.

The 'en' language is the only language which is supported because the model knows no other languages.

For now just a field in the database will be set, no visual hint if a message is spam or not.